Move the CtoScala Converter to an individual place
